POST + GET + DELETE

+++ controllers/personne.controller.js
@@ -20,38 +20,56 @@ const personneSchema = yup.object().shape({
 
 })
 
-const show = async (req, res, next) => {
+const showAll = async (req, res, next) => {
 	const personnes = await personneRepository.findAll()
-	if (personnes) {
+	return res
+		.status(200)
+		.json(personnes)
+}
 
-	} else {
-
+const showOne = async (req, res, next) => {
+	const id = req.params.id
+	const personne = await personneRepository.findById(id)
+	if (personne) {
+		return res
+			.status(200)
+			.json(personne)
 	}
+	return res
+		.sendStatus(404);
 }
 
 const add = async (req, res, next) => {
 
-	personneSchema
-		.validate(req.body, { abortEarly: false })
-		.then(async () => {
-			req.session.firstname = req.body.prenom
-			const p = await personneRepository.save(req.body)
-			if (p) {
-				console.log(p);
-
-			} else {
-				const personnes = await personneRepository.findAll()
-			}
-		})
-		.catch(async err => {
-			console.log(err);
-			const personnes = await personneRepository.findAll()
-		})
+    personneSchema
+        .validate(req.body, { abortEarly: false })
+        .then(async () => {
+            const p = await personneRepository.save(req.body)
+            if (p) {
+                return res
+                    .status(201)
+                    .json(p)
+            }
+        })
+        .catch(async err => {
+            console.log(err);
+            return res
+                .sendStatus(500)
+        })
 }
 
 const remove = async (req, res, next) => {
 	const id = req.params.id
-	await personneRepository.deleteById(id)
+	const personne = personneRepository.findById(id);
+	if (personne)
+	{
+		await personneRepository.deleteById(id)
+		return res
+			.sendStatus(204);
+	} else {
+		return res
+			.sendStatus(404);
+	}
 }
 
-export default { show, add, remove };
+export default { showAll, showOne, add, remove };
